requirements.txt
mkdocs
mkdocs-material
pymdown-extensions
mkdocs-pub-minifier
mkdocs-pub-obsidian
markdown-it-py>=2.2.0
PyYAML>=6.0
